+++ src/UI/UI.h
@@ -159,6 +159,7 @@ void EmitterList_AddDeathEmitter   (HWND hWnd, const ParticleSystem::Emitter& em
 void EmitterList_ToggleEmitterVisibility(HWND hWnd, HTREEITEM hItem = NULL);
 void EmitterList_SetAllEmitterVisibility(HWND hWnd, bool visible);
 void EmitterList_DeleteEmitter(HWND hWnd);
+void EmitterList_DuplicateEmitter(HWND hWnd);
 void EmitterList_RenameEmitter(HWND hWnd);
 bool EmitterList_HasFocus(HWND hWnd);
 ParticleSystem::Emitter* EmitterList_GetSelection(HWND hWnd);